proto/dbft: use uint64 for block numbers
uint32 is too small for Ethereum, it's native block number size is 64-bit.

+++ eth/protocols/dbft/ledger.go
@@ -16,8 +16,8 @@ func newLedger(bc BlockChainAPI) *ledger {
 	return &ledger{bc: bc}
 }
 
-func (l *ledger) BlockHeight() uint32 {
-	return uint32(l.bc.BlockNumber())
+func (l *ledger) BlockHeight() uint64 {
+	return uint64(l.bc.BlockNumber())
 }
 
 func (l *ledger) IsAddressAllowed(addr common.Address) bool {
